Move converter code to its own file
