+++ packages/fast-web-utilities/src/index.ts
@@ -7,3 +7,4 @@ export * from "./localization";
 export * from "./numbers";
 export * from "./strings";
 export * from "./query";
+export * from "./rtl-scroll-converter";
